Perf update

+++ grails-app/controllers/com/mogobiz/store/partner/ImpexController.groovy
@@ -7,14 +7,10 @@ import com.mogobiz.service.ImportService
 import com.mogobiz.store.domain.Catalog
 import com.mogobiz.store.domain.Company
 import com.mogobiz.store.domain.Seller
-import com.sun.corba.se.spi.orbutil.threadpool.Work
 import grails.converters.JSON
 import grails.transaction.Transactional
-import groovy.sql.Sql
 import org.hibernate.SessionFactory
 
-import java.sql.Connection
-import java.sql.SQLException
 import java.util.zip.ZipFile
 
 class ImpexController {
@@ -65,8 +61,6 @@ class ImpexController {
 
     @Transactional
     def ximport() {
-        log.info("IMPORT STARTED")
-        Date start = new Date()
         def seller = request.seller ? request.seller : authenticationService.retrieveAuthenticatedSeller()
         if (!seller) {
             response.sendError 401
@@ -75,15 +69,22 @@ class ImpexController {
         log.info("Uploading file ...")
         def file = request.getFile('file')
         if (file && !file.empty) {
+            log.info("IMPORT STARTED")
+            Date start = new Date()
             File tmpFile = File.createTempFile("import", ".zip")
             file.transferTo(tmpFile)
             Company company = seller.company
             def name = "impex"
-            Catalog catalog = Catalog.findByNameAndCompany(name, seller.company)
             int countSales = 0
-            if (catalog)
-                countSales = catalogService.purge(catalog.id)
-
+            Catalog catalog
+            Catalog.withNewTransaction {
+                catalog = Catalog.findByNameAndCompany(name, seller.company)
+                if (catalog) {
+                    log.info("Purging catalog ...")
+                    countSales = catalogService.purge(catalog.id)
+                    log.info("Purge ended ...")
+                }
+            }
             if (countSales == 0) {
                 catalog = new Catalog()
                 catalog.company = company
@@ -91,11 +92,17 @@ class ImpexController {
                 catalog.activationDate = new Date(2040 - 1900, 11, 31)
                 catalog.uuid = UUID.randomUUID().toString()
                 if (catalog.validate()) {
-                    catalog.save(flush: true)
-                    importService.ximport(catalog, new ZipFile(tmpFile))
+                    Catalog.withNewTransaction {
+                        catalog.save(flush: true)
+                    }
                 } else {
                     System.out.println(catalog.errors)
+                    catalog = null
                 }
+            }
+
+            if (countSales == 0 && catalog) {
+                importService.ximport(catalog, new ZipFile(tmpFile))
                 tmpFile.delete()
                 withFormat {
                     json { render catalog as JSON }
@@ -104,11 +111,11 @@ class ImpexController {
                 tmpFile.delete()
                 response.sendError(403, "$countSales")
             }
+            log.info("IMPORT FINISHED")
+            Date end = new Date()
+            log.info("IMPORT DURATION (in seconds) =" + (end.getTime() - start.getTime()) / 1000)
         } else {
             response.sendError(401, "Missing file")
         }
-        log.info("IMPORT FINISHED")
-        Date end = new Date()
-        println("IMPORT DURATION (in seconds) =" + (end.getTime() - start.getTime()) / 1000)
     }
 }
